replaced deprecated modules, added mozilla agent to url request, and fixed some tensors and numbers that was supposed to be ints

+++ models/decoders/attention_scn.py
@@ -249,7 +249,7 @@ class AttentionSCN(nn.Module):
                     -1).topk(k, 0, True, True)
 
             # Convert unrolled indices to actual indices of scores
-            prev_word_inds = top_k_words / vocab_size  # (s)
+            prev_word_inds = (top_k_words / vocab_size).long()  # (s)
             next_word_inds = top_k_words % vocab_size  # (s)
 
             # Add new words to sequences, alphas
